Issue #21 Split out a standard HTTP runner and start an Undertow server.
