+++ unit/auth-authz/pom.xml
@@ -45,6 +45,16 @@
   <dependencies>
     <!-- Testsuite Module Dependencies -->
     <dependency>
+      <groupId>io.undertow</groupId>
+      <artifactId>undertow-core</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.undertow</groupId>
+      <artifactId>undertow-servlet</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
       <groupId>org.wildfly.security.testsuite</groupId>
       <artifactId>wildfly-security-common-auth-authz</artifactId>
     </dependency>
