+++ unit/auth-authz/src/test/java/org/wildfly/security/tests/authauthz/DynamicAuthPermutationsSuiteRunner.java
@@ -80,31 +80,6 @@ public class DynamicAuthPermutationsSuiteRunner {
     }
 
     @TestFactory
-    Stream<DynamicTest> dynamicHttpTests() {
-        System.out.println("DynamicAuthPermutationsSuiteRunner->dynamicHttpTests");
-        List<DynamicTest> dynamicTests = new ArrayList<>();
-
-        TestContext testContext = getTestContext();
-        if (testContext != null && testContext.enabledTransports().contains(Transport.HTTP)) {
-            final String mode = getMode();
-            testContext.mechanismsForTransport(Transport.HTTP).forEach(
-                    s -> {
-                        dynamicTests.add(dynamicTest(String.format("[%s] testHttpSuccess(%s)", mode, s),
-                                () -> testHttpSuccess(s)));
-                        dynamicTests.add(dynamicTest(String.format("[%s] testHttpBadUsername(%s)", mode, s),
-                                () -> testHttpBadUsername(s)));
-                        dynamicTests.add(dynamicTest(String.format("[%s] testHttpBadPassword(%s)", mode, s),
-                                () -> testHttpBadPassword(s)));
-                        dynamicTests.add(dynamicTest(String.format("[%s] testHttpBruteForce(%s)", mode, s),
-                                () -> testHttpBruteForce(s)));
-                    }
-            );
-        }
-
-        return dynamicTests.stream();
-    }
-
-    @TestFactory
     Stream<DynamicTest> dynamicTests() {
         System.out.println("DynamicAuthPermutationsSuiteRunner->dynamicTests");
         return Stream.of(Transport.values())
@@ -167,18 +142,6 @@ public class DynamicAuthPermutationsSuiteRunner {
 
     public void testSaslBruteForce(final String mechanism) {
         System.out.printf("testSaslBruteForce(%s)\n", mechanism);
-    }
-
-    public void testHttpSuccess(final String mechanism) {
-        System.out.printf("testHttpSuccess(%s)\n", mechanism);
-    }
-
-    public void testHttpBadUsername(final String mechanism) {
-        System.out.printf("testHttpBadUsername(%s)\n", mechanism);
-    }
-
-    public void testHttpBadPassword(final String mechanism) {
-        System.out.printf("testHttpBadPassword(%s)\n", mechanism);
     }
 
     public void testHttpBruteForce(final String mechanism) {
